Integrate home page
Remove insurance contents.

Put `AboutUs` and `Partners` on the home page.
